Add files via upload
I developed this code using python, the (Face, Age and Eye Detection) software. I have made this for a project I have been working on The Smart-Helmet so as to detect5 the face and eye position mainly of a person while he would be driving. This software can also be used in many other security purposes needed in military or government purposes.
You are welcome to please come and contribute to this project.